fix(indexer): address v3 test harness review
